requirements-dev.txt
requests>=2.31.0
fastapi
httpx
PyGithub>=2.3.0
jsonschema
uvicorn
pytest
